Added many Lidl and Kaufland
Added many Lidl and Kaufland in Germany, Austria, Romania, Poland and Hungary. Source for Germany is the Lidl Plus App combined with Google Maps, for the other it's mostly Plugshare and checked in Google Maps. Austria is from an address list from the Lidl Austria Website checked with Google Maps. On the first look it looks like I deleted something, but I just sorted my new chargers with the old ones and the old list was partly not sorted correctly. At example Lidl Baden-Baden was after Lidel Berlin.